pyproject.toml
[tool.poetry]
name = "xlearn"
version = "0.1.0"
description = ""
authors = ["ShinnosukeUesaka <shinnosuke.uesaka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
xai-sdk = "*"
python-dotenv = "^1.0.1"
tweepy = "^4.14.0"
fastapi = "^0.110.2"
oauth2 = "^1.9.0.post1"
uvicorn = "^0.29.0"
jinja2 = "^3.1.3"
firebase-admin = "^6.5.0"
openai = "^1.23.2"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
